+++ frontend/src/routes/_layout.tsx
@@ -1,4 +1,5 @@
-import { createFileRoute, Outlet, redirect } from "@tanstack/react-router"
+import { createFileRoute, Outlet } from "@tanstack/react-router"
+// import { isLoggedIn } from "@/hooks/useAuth"
 
 import { Footer } from "@/components/Common/Footer"
 import AppSidebar from "@/components/Sidebar/AppSidebar"
@@ -7,17 +8,16 @@ import {
   SidebarProvider,
   SidebarTrigger,
 } from "@/components/ui/sidebar"
-import { isLoggedIn } from "@/hooks/useAuth"
 
 export const Route = createFileRoute("/_layout")({
   component: Layout,
-  beforeLoad: async () => {
-    if (!isLoggedIn()) {
-      throw redirect({
-        to: "/login",
-      })
-    }
-  },
+  //   beforeLoad: async () => {
+  //     if (!isLoggedIn()) {
+  //       throw redirect({
+  //         to: "/login",
+  //       })
+  //     }
+  //   },
 })
 
 function Layout() {